+++ RELEASE-NOTES.md
@@ -48,6 +48,7 @@ END TEMPLATE-->
 * Fix `Menu` and `NumpadDecimal` key codes on SDL3.
 * Fix modifying `Label.FontOverride` not causing a layout update.
 * Controls created by rich-text tags now get arranged to a proper size.
+* Fix `OutputPanel` scrollbar breaking if a style update changes the font size.
 
 ### Other